Clean up Obj-C stubbing and verification files
