+++ Sources/MockingbirdFramework/Stubbing/StubbingContext.swift
@@ -53,80 +53,6 @@ import XCTest
       .implementationProvider()
   }
   
-  @objc public static let noImplementation = NSObject()
-  
-  @objc public func returnValue(for invocation: ObjCInvocation) -> Any? {
-    let implementation = implementation(for: invocation as Invocation)
-    if let concreteImplementation = implementation as? () -> Any? {
-      return concreteImplementation()
-    } else if let concreteImplementation = implementation
-                as? (Any?) -> Any? {
-      return concreteImplementation(invocation.arguments.get(0)?.base)
-    } else if let concreteImplementation = implementation
-                as? (Any?, Any?) -> Any? {
-      return concreteImplementation(invocation.arguments.get(0)?.base,
-                                    invocation.arguments.get(1)?.base)
-    } else if let concreteImplementation = implementation
-                as? (Any?, Any?, Any?) -> Any? {
-      return concreteImplementation(invocation.arguments.get(0)?.base,
-                                    invocation.arguments.get(1)?.base,
-                                    invocation.arguments.get(2)?.base)
-    } else if let concreteImplementation = implementation
-                as? (Any?, Any?, Any?, Any?) -> Any? {
-      return concreteImplementation(invocation.arguments.get(0)?.base,
-                                    invocation.arguments.get(1)?.base,
-                                    invocation.arguments.get(2)?.base,
-                                    invocation.arguments.get(3)?.base)
-    } else if let concreteImplementation = implementation
-                as? (Any?, Any?, Any?, Any?, Any?) -> Any? {
-      return concreteImplementation(invocation.arguments.get(0)?.base,
-                                    invocation.arguments.get(1)?.base,
-                                    invocation.arguments.get(2)?.base,
-                                    invocation.arguments.get(3)?.base,
-                                    invocation.arguments.get(4)?.base)
-    } else if let concreteImplementation = implementation
-                as? (Any?, Any?, Any?, Any?, Any?, Any?) -> Any? {
-      return concreteImplementation(invocation.arguments.get(0)?.base,
-                                    invocation.arguments.get(1)?.base,
-                                    invocation.arguments.get(2)?.base,
-                                    invocation.arguments.get(3)?.base,
-                                    invocation.arguments.get(4)?.base,
-                                    invocation.arguments.get(5)?.base)
-    } else if let concreteImplementation = implementation
-                as? (Any?, Any?, Any?, Any?, Any?, Any?, Any?) -> Any? {
-      return concreteImplementation(invocation.arguments.get(0)?.base,
-                                    invocation.arguments.get(1)?.base,
-                                    invocation.arguments.get(2)?.base,
-                                    invocation.arguments.get(3)?.base,
-                                    invocation.arguments.get(4)?.base,
-                                    invocation.arguments.get(5)?.base,
-                                    invocation.arguments.get(6)?.base)
-    } else if let concreteImplementation = implementation
-                as? (Any?, Any?, Any?, Any?, Any?, Any?, Any?, Any?) -> Any? {
-      return concreteImplementation(invocation.arguments.get(0)?.base,
-                                    invocation.arguments.get(1)?.base,
-                                    invocation.arguments.get(2)?.base,
-                                    invocation.arguments.get(3)?.base,
-                                    invocation.arguments.get(4)?.base,
-                                    invocation.arguments.get(5)?.base,
-                                    invocation.arguments.get(6)?.base,
-                                    invocation.arguments.get(7)?.base)
-    } else if let concreteImplementation = implementation
-                as? (Any?, Any?, Any?, Any?, Any?, Any?, Any?, Any?, Any?) -> Any? {
-      return concreteImplementation(invocation.arguments.get(0)?.base,
-                                    invocation.arguments.get(1)?.base,
-                                    invocation.arguments.get(2)?.base,
-                                    invocation.arguments.get(3)?.base,
-                                    invocation.arguments.get(4)?.base,
-                                    invocation.arguments.get(5)?.base,
-                                    invocation.arguments.get(6)?.base,
-                                    invocation.arguments.get(7)?.base,
-                                    invocation.arguments.get(8)?.base)
-    }
-    
-    return Self.noImplementation
-  }
-
   func clearStubs() {
     stubs.update { $0.removeAll() }
   }
